# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "panda-wiki API documentation",
        "title": "panda-wiki API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/node/detail": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Get Node Detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "node"
                ],
                "summary": "Get Node Detail",
                "parameters": [
                    {
                        "type": "string",
                        "name": "format",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "kb_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.PWResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/v1.NodeDetailResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Update Node Detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "node"
                ],
                "summary": "Update Node Detail",
                "parameters": [
                    {
                        "description": "Node",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateNodeReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/node/list": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Get Node List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "node"
                ],
                "summary": "Get Node List",
                "parameters": [
                    {
                        "type": "string",
                        "name": "kb_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.PWResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.NodeListItemResp"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/share/v1/node/detail": {
            "get": {
                "description": "GetNodeDetail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "share_node"
                ],
                "summary": "GetNodeDetail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "kb id",
                        "name": "X-KB-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "node id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "format",
                        "name": "format",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    }
                }
            }
        },
        "/share/v1/node/list": {
            "get": {
                "description": "GetNodeList",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "share_node"
                ],
                "summary": "GetNodeList",
                "parameters": [
                    {
                        "type": "string",
                        "description": "kb id",
                        "name": "X-KB-ID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    }
                }
            }
        },
        "/share/v1/stat/page": {
            "post": {
                "description": "RecordPage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "share_stat"
                ],
                "summary": "RecordPage",
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.StatPageReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "consts.NodeAccessPerm": {
            "type": "string",
            "enum": [
                "open",
                "partial",
                "closed"
            ],
            "x-enum-comments": {
                "NodeAccessPermClosed": "完全禁止",
                "NodeAccessPermOpen": "完全开放",
                "NodeAccessPermPartial": "部分开放"
            },
            "x-enum-descriptions": [
                "完全开放",
                "部分开放",
                "完全禁止"
            ],
            "x-enum-varnames": [
                "NodeAccessPermOpen",
                "NodeAccessPermPartial",
                "NodeAccessPermClosed"
            ]
        },
...
        "domain.NodeListItemResp": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "creator": {
                    "type": "string"
                },
                "creator_id": {
                    "type": "string"
                },
                "editor": {
                    "type": "string"
                },
                "editor_id": {
                    "type": "string"
                },
                "emoji": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string"
                },
                "permissions": {
                    "$ref": "#/definitions/domain.NodePermissions"
                },
                "position": {
                    "type": "number"
                },
                "status": {
                    "$ref": "#/definitions/domain.NodeStatus"
                },
                "summary": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/domain.NodeType"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.NodeMeta": {
            "type": "object",
            "properties": {
                "emoji": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                }
            }
        },
        "domain.NodePermissions": {
            "type": "object",
            "properties": {
                "answerable": {
                    "description": "可被问答",
                    "allOf": [
                        {
                            "$ref": "#/definitions/consts.NodeAccessPerm"
                        }
                    ]
                },
                "visible": {
                    "description": "导航内可见",
                    "allOf": [
                        {
                            "$ref": "#/definitions/consts.NodeAccessPerm"
                        }
                    ]
                },
                "visitable": {
                    "description": "可被访问",
                    "allOf": [
                        {
                            "$ref": "#/definitions/consts.NodeAccessPerm"
                        }
                    ]
                }
            }
        },
        "domain.NodeStatus": {
            "type": "integer",
            "format": "int32",
            "enum": [
                1,
                2
            ],
            "x-enum-varnames": [
                "NodeStatusDraft",
                "NodeStatusReleased"
            ]
        },
...
        "domain.NodeType": {
            "type": "integer",
            "format": "int32",
            "enum": [
                1,
                2
            ],
            "x-enum-varnames": [
                "NodeTypeFolder",
                "NodeTypeDocument"
            ]
        },
...
        "domain.PWResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "domain.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "domain.StatPageReq": {
            "type": "object",
            "required": [
                "scene"
            ],
            "properties": {
                "node_id": {
                    "type": "string"
                },
                "scene": {
                    "enum": [
                        1,
                        2,
                        3,
                        4
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.StatPageScene"
                        }
                    ]
                }
            }
        },
        "domain.StatPageScene": {
            "type": "integer",
            "enum": [
                1,
                2,
                3,
                4
            ],
            "x-enum-varnames": [
                "StatPageSceneWelcome",
                "StatPageSceneNodeDetail",
                "StatPageSceneChat",
                "StatPageSceneLogin"
            ]
        },
...
        "domain.UpdateNodeReq": {
            "type": "object",
            "required": [
                "id",
                "kb_id"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "emoji": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "kb_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "position": {
                    "type": "number"
                },
                "summary": {
                    "type": "string"
                }
            }
        },
...
        "v1.NodeDetailResp": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "kb_id": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/domain.NodeMeta"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string"
                },
                "permissions": {
                    "$ref": "#/definitions/domain.NodePermissions"
                },
                "status": {
                    "$ref": "#/definitions/domain.NodeStatus"
                },
                "type": {
                    "$ref": "#/definitions/domain.NodeType"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "bearerAuth": {
            "description": "Type \"Bearer\" + a space + your token to authorize",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}